RavenDB-19021 : initial work
