+++ src/Raven.Server/Documents/PeriodicBackup/Restore/RestoreBackupTask.cs
@@ -2,7 +2,6 @@
 using System.Threading.Tasks;
 using Raven.Client.Documents.Operations.Backups;
 using Raven.Server.ServerWide;
-using Raven.Server.ServerWide.Context;
 
 namespace Raven.Server.Documents.PeriodicBackup.Restore
 {
