+++ src/Raven.Server/Documents/PeriodicBackup/Restore/Sharding/SingleShardRestoreBackupTask.cs
@@ -1,15 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Diagnostics;
 using System.Threading.Tasks;
-using Raven.Client.Documents.Operations;
 using Raven.Client.Documents.Operations.Backups;
+using Raven.Client.Documents.Smuggler;
 using Raven.Client.ServerWide;
-using Raven.Client.ServerWide.Operations;
 using Raven.Server.Config;
 using Raven.Server.ServerWide;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.Smuggler.Documents;
+using Raven.Server.Smuggler.Documents.Data;
 using Raven.Server.Utils;
 using Sparrow.Json;
 using Sparrow.Utils;
@@ -18,7 +17,7 @@ namespace Raven.Server.Documents.PeriodicBackup.Restore.Sharding
 {
     internal class SingleShardRestoreBackupTask : RestoreBackupTask
     {
-        private readonly string _nonShardedDbName;
+        private readonly int _shardNumber;
 
         public SingleShardRestoreBackupTask(ServerStore serverStore, RestoreBackupConfigurationBase restoreConfiguration, List<string> filesToRestore, 
             IRestoreSource restoreSource, OperationCancelToken operationCancelToken) : base(serverStore, restoreConfiguration, restoreSource, filesToRestore, operationCancelToken)
@@ -27,13 +26,25 @@ namespace Raven.Server.Documents.PeriodicBackup.Restore.Sharding
 
             DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Aviv, DevelopmentHelper.Severity.Major, "test encrypted backup & restore");
             HasEncryptionKey = false;
+
+            _shardNumber = ShardHelper.GetShardNumber(DatabaseName);
+        }
+        
+        protected override Task OnBeforeRestoreAsync()
+        {
+            CreateDocumentDatabase();
             
-            _nonShardedDbName = ShardHelper.ToDatabaseName(DatabaseName);
+            var topology = RestoreSettings.DatabaseRecord.Sharding.Shards[_shardNumber];
+            Database.SetIds(topology, RestoreSettings.DatabaseRecord.Sharding.DatabaseId);
+
+            return Task.CompletedTask;
         }
 
-        protected override RavenConfiguration CreateDatabaseConfiguration()
+        protected override Task OnAfterRestoreAsync()
         {
-            return DatabasesLandlord.CreateDatabaseConfiguration(ServerStore, DatabaseName, RestoreSettings.DatabaseRecord.Settings);
+            SmugglerBase.EnsureProcessed(Result, skipped: false);
+            Progress.Invoke(Result.Progress);
+            return Task.CompletedTask;
         }
 
         protected override DatabaseRecord GetDatabaseRecord()
@@ -41,28 +52,25 @@ namespace Raven.Server.Documents.PeriodicBackup.Restore.Sharding
             using (ServerStore.ContextPool.AllocateOperationContext(out TransactionOperationContext serverContext))
             using (serverContext.OpenReadTransaction())
             {
-                var databaseRecord = ServerStore.Cluster.ReadDatabase(serverContext, _nonShardedDbName);
+                var databaseRecord = ServerStore.Cluster.ReadDatabase(serverContext, ShardHelper.ToDatabaseName(DatabaseName));
                 Debug.Assert(databaseRecord != null);
 
                 return databaseRecord;
             }
         }
 
-        protected override Task OnBeforeRestoreAsync()
+        protected override RavenConfiguration CreateDatabaseConfiguration()
         {
-            CreateDocumentDatabase();
-            return Task.CompletedTask;
+            return DatabasesLandlord.CreateDatabaseConfiguration(ServerStore, DatabaseName, RestoreSettings.DatabaseRecord.Settings);
         }
 
-        protected override void OnAfterRestore()
+        protected override async Task ImportLastBackupFileAsync(DocumentDatabase database, DatabaseDestination destination, JsonOperationContext context,
+            DatabaseSmugglerOptionsServerSide options, DatabaseRecord databaseRecord, string lastFilePath)
         {
-            SmugglerBase.EnsureProcessed(Result, skipped: false);
-            Progress.Invoke(Result.Progress);
-        }
+            if (_shardNumber > 0)
+                options.OperateOnTypes &= ~DatabaseItemType.Subscriptions;
 
-        protected override Task<long> SaveDatabaseRecordAsync(string databaseName, DatabaseRecord databaseRecord, Dictionary<string, BlittableJsonReaderObject> databaseValues, RestoreResult restoreResult, Action<IOperationProgress> onProgress)
-        {
-            return base.SaveDatabaseRecordAsync(_nonShardedDbName, databaseRecord, databaseValues, restoreResult, onProgress);
+            await ImportSingleBackupFileAsync(database, Progress, Result, lastFilePath, context, destination, options, isLastFile: true);
         }
     }
 }
